+++ src/main/java/kitchenpos/eatinorders/tobe/domain/constant/EatInOrderStatus.java
@@ -1,4 +1,4 @@
-package kitchenpos.eatinorders.tobe.constant;
+package kitchenpos.eatinorders.tobe.domain.constant;
 
 public enum EatInOrderStatus {
     WAITING, ACCEPTED, SERVED, COMPLETED
