+++ src/main/java/kitchenpos/eatinorders/tobe/domain/constant/EatInOrderType.java
@@ -1,0 +1,5 @@
+package kitchenpos.eatinorders.tobe.domain.constant;
+
+public enum EatInOrderType {
+    EAT_IN
+}
