+++ src/main/java/kitchenpos/eatinorders/tobe/domain/repository/OrderTableRepository.java
@@ -1,6 +1,6 @@
-package kitchenpos.eatinorders.tobe.repository;
+package kitchenpos.eatinorders.tobe.domain.repository;
 
-import kitchenpos.eatinorders.tobe.entity.OrderTable;
+import kitchenpos.eatinorders.tobe.domain.entity.OrderTable;
 
 import java.util.Optional;
 import java.util.UUID;
